% Generated by roxygen2 (4.0.1): do not edit by hand
\name{clock_dataframe}
\alias{clock_dataframe}
\title{Aggregate the org elements of an org file in a data.frame.}
\usage{
clock_dataframe(x, units = "mins", inherit_tags = TRUE,
  pattern_states = "\\\\b[[:upper:]]{2,}((\\\\b)|(_[[:upper:]]+\\\\b))(?!(\\\\s{2,}|\\\\t|$))",
  pattern_efforts = "^[ :]+Effort:")
}
\arguments{
\item{x}{org object as character vector.}

\item{units}{unit of time used in the columns Effort, Time_Spent
and Period.}

\item{inherit_tags}{logical; if \code{TRUE} (default), inherit
tags from level one headlines.}

\item{pattern_states}{a regular expression to match the
\code{org-todo-keywords} in the headings.}

\item{pattern_efforts}{a regular expression to match the effort
entries.}
}
\value{
data.frame object.
}
\description{
This function creates a \code{data.frame} object that aggregates
the org elements of each heading in a row. In addition,
\code{First} and \code{Last} reveal the first and the last
timestamp stored for each heading. The period between \code{First}
and \code{Last} is returned in the \code{Period} column. The unit
of time in the columns Effort, Time_Spent and Period is controlled
by the \code{units} parameter. Efforts make use of the time units
\code{"mins"} and \code{"hours"} only, because the format H:MM is
hard-coded for the effort property in org-mode. If the
\code{inherit_tags} parameter is set to \code{TRUE}, the tags from
level one headings will be inherited. Use \code{pattern_states}
and \code{pattern_efforts} to control the regular expressions to
match the \code{org-todo-keywords} and the effort entries in an
org file.
}
\examples{
system.file("extdata", "sample.org", package = "orgclockr") \%>\%
readLines() \%>\%
clock_dataframe(units = "hours") \%>\%
head(3)
##      Category           Tag   Headline Level State Effort Time_Spent
## 1 CategoryOne        TagOne HeadingOne     1  <NA>   <NA>       0.00
## 2        <NA> TagOne TagTwo    TaskOne     2  TODO   <NA>       0.00
## 3        <NA>        TagOne    TaskTwo     2  TODO   0.33       0.35
##                 First                Last Period
## 1                <NA>                <NA>     NA
## 2                <NA>                <NA>     NA
## 3 2014-10-08 Mi 13:34 2014-10-10 Fr 17:05  51.52
}
\seealso{
\code{extract_time_spent}, \code{extract_categories},
\code{extract_todostates}, \code{extract_headlines},
\code{extract_timestamps}, \code{extract_tags},
\code{extract_efforts} and \code{extract_levels} to extract
particular org elements.
}

