% Generated by roxygen2 (4.0.1): do not edit by hand
\name{extract_categories_total}
\alias{extract_categories_total}
\title{Extract the categories of an org file.}
\usage{
extract_categories_total(x)
}
\arguments{
\item{x}{org object as character vector.}
}
\value{
the categories of an orgfile as a character vector.
}
\description{
Extract the categories of an org file.
}

