% Generated by roxygen2 (4.0.1): do not edit by hand
\name{extract_days_on_task}
\alias{extract_days_on_task}
\title{Extract the number of days a task has been worked on.}
\usage{
extract_days_on_task(x)
}
\arguments{
\item{x}{}
}
\value{
the days on a task in an org file as a numeric vector.
}
\description{
Do not confuse the number of days on a task with the total time
spent on a task in days. The latter can be calculated from
\\code{extract_time_spent}.
}
\examples{
system.file("extdata", "sample.org", package = "orgclockr") \%>\%
readLines() \%>\%
extract_days_on_task()
##  [1] 0 0 2 0 0 1 5 0 1 2 1 2
}
\seealso{
\code{extract_days_on_task}, \code{extract_time_spent},
\code{extract_categories}, \code{extract_todostates},
\code{extract_headlines}, \code{extract_timestamps},
\code{extract_tags}, \code{extract_efforts} and
\code{extract_levels} to extract other org elements.
}

