% Generated by roxygen2 (4.0.1): do not edit by hand
\name{extract_deadlines}
\alias{extract_deadlines}
\title{Extract deadlines of headings in an org file.}
\usage{
extract_deadlines(x, pattern = "^\\\\ {0,}DEADLINE:\\\\ +<[[:digit:]].*>$")
}
\arguments{
\item{x}{org object as character vector.}

\item{pattern}{a regular expression to match the deadlines set in
the org file.}
}
\value{
the deadlines of an org file as a character vector.
}
\description{
This function extracts the timestamps in lines that match the
\code{pattern}. According to \code{org-deadline-string} these
lines should start with "DEADLINE:".
}
\examples{
system.file("extdata", "sample.org", package = "orgclockr") \%>\%
readLines() \%>\%
extract_deadlines()
##  [1] NA              NA              NA              NA
##  [5] NA              "2014-01-04 Sa" NA              NA
##  [9] NA              NA              NA              NA
}
\seealso{
\code{extract_intervals}, \code{extract_days_on_task},
\code{extract_time_spent}, \code{extract_categories},
\code{extract_todostates}, \code{extract_headlines},
\code{extract_timestamps}, \code{extract_tags} and
\code{extract_levels} to extract other org elements.
}

