% Generated by roxygen2 (4.0.1): do not edit by hand
\name{extract_efforts}
\alias{extract_efforts}
\title{Extract the efforts set in an org file.}
\usage{
extract_efforts(x, pattern = "^[ :]+Effort:", units = "mins")
}
\arguments{
\item{pattern}{a regular expression to match the property that is
being used to keep track of effort estimates.}

\item{units}{unit of time used. \code{mins} and \code{hours} are
valid options.}

\item{x}{org object as character vector.}
}
\value{
the efforts set in an org file as a numeric vector.
These have the format H:MM.
}
\description{
This function extracts the efforts matching the regexp pattern.
The pattern matches the property \code{Effort} according to the
settings in \code{org-effort-property} by default. The format H:MM
for this property is hard-coded, therefore this function will use
either minutes or hours as units of time in the resulting vector.
}
\examples{
system.file("extdata", "sample.org", package = "orgclockr") \%>\%
readLines() \%>\%
extract_efforts()
##  [1]  NA  NA  20  NA  NA  NA  60  NA  30  25 240  25
}
\seealso{
\code{extract_intervals}, \code{extract_days_on_task},
\code{extract_time_spent}, \code{extract_categories},
\code{extract_todostates}, \code{extract_headlines},
\code{extract_timestamps}, \code{extract_tags} and
\code{extract_levels} to extract other org elements.
}

