% Generated by roxygen2 (4.0.1): do not edit by hand
\name{extract_efforts_total}
\alias{extract_efforts_total}
\title{Extract all efforts set in an org file.}
\usage{
extract_efforts_total(x, pattern)
}
\arguments{
\item{x}{org object as character vector.}

\item{pattern}{regular expression to match the effort property set
by \code{org-effort-property}.}
}
\value{
all efforts set in an org file.
}
\description{
The regular expression to match the effort property in an org file
will be passed from \code{extract_efforts}.
}

