% Generated by roxygen2 (4.0.1): do not edit by hand
\name{extract_headlines}
\alias{extract_headlines}
\title{Extract the headlines of an org file.}
\usage{
extract_headlines(orgfile)
}
\arguments{
\item{orgfile}{org object as character vector.}
}
\value{
the headlines of an orgfile as a character matrix.
}
\description{
This function is used to extract the headlines matching the regexp
pattern.
}
\seealso{
\code{extract_tags}, \code{extract_timestamps}
and \code{extract_categories} for extractions of other org elements.
}

