% Generated by roxygen2 (4.0.1): do not edit by hand
\name{extract_intervals}
\alias{extract_intervals}
\title{Extract clock intervals of headings in an org file.}
\usage{
extract_intervals(x, units = "mins")
}
\arguments{
\item{x}{org object as character vector.}

\item{units}{unit of time used.}
}
\value{
a list of double values in the unit of time given.
}
\description{
This expects an even number of timestamps per headline, so in org
files with running clocks the lonely timestamp will be ignored.
}
\details{
Org mode uses the standard ISO notation for dates and times as it
is defined in ISO 8601. The format is set in
\code{org-time-stamp-custom-formats}. If this variable is
modified, the extraction is likely to fail.
}
\examples{
system.file("extdata", "sample.org", package = "orgclockr") \%>\%
readLines() \%>\%
extract_intervals() \%>\%
last()
## [1]  68  17   1  62  16 170
}

