% Generated by roxygen2 (4.0.1): do not edit by hand
\name{extract_raw_headlines}
\alias{extract_raw_headlines}
\title{Extract the raw headlines of an org file.}
\usage{
extract_raw_headlines(x)
}
\arguments{
\item{x}{org object as character vector.}
}
\value{
headlines as the whole lines.
}
\description{
Extract the raw headlines of an org file.
}

