% Generated by roxygen2 (4.0.1): do not edit by hand
\name{extract_tags}
\alias{extract_tags}
\title{Extract the tags of an org file.}
\usage{
extract_tags(x)
}
\arguments{
\item{x}{org object as character vector.}
}
\value{
the tags of an orgfile as a character matrix.
}
\description{
This function is used to extract the tags matching the regexp
pattern.
}
\seealso{
\code{extract_headlines}, \code{extract_timestamps},
\code{extract_categories} and \code{extract_levels} for
extractions of other org elements.
}

