% Generated by roxygen2 (4.0.1): do not edit by hand
\name{extract_timestamps_total}
\alias{extract_timestamps_total}
\title{Extract the timestamps of an org file.}
\usage{
extract_timestamps_total(x)
}
\arguments{
\item{x}{org object as character vector.}
}
\value{
the timestamps of an orgfile as a list.
}
\description{
Extract the timestamps of an org file.
}

