% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{orgfile}
\alias{orgfile}
\title{A typical org file.}
\format{A character vector with 96 variables.}
\usage{
orgfile
}
\description{
A character object resulting from reading in an org file of nearly
100 lines. For presentational purposes the file was enriched with
several elements typical for todo items in Org mode.
}
\keyword{datasets}

