% Generated by roxygen2 (4.0.1): do not edit by hand
\name{split_orgfile}
\alias{split_orgfile}
\title{Split the org file by nodes.}
\usage{
split_orgfile(x)
}
\arguments{
\item{x}{org object as character vector.}
}
\value{
the nodes of an org file as a character list.
}
\description{
\code{split_orgfile} splits the org file by the index of the
headlines.
}

