% Generated by roxygen2 (4.0.1): do not edit by hand
\name{extract_categories}
\alias{extract_categories}
\title{Extract the categories of an org file.}
\usage{
extract_categories(x)
}
\arguments{
\item{x}{org object as character vector.}
}
\value{
the categories of an orgfile as a character matrix.
}
\description{
This function is used to extract the categories matching the
regexp pattern.
}
\seealso{
\code{extract_headlines}, \code{extract_timestamps},
\code{extract_tags} and \code{extract_levels} for extractions of
other org elements.
}

