% Generated by roxygen2 (4.0.1): do not edit by hand
\name{extract_days_on_task}
\alias{extract_days_on_task}
\title{Extract the number of days a task has been worked on.}
\usage{
extract_days_on_task(x)
}
\arguments{
\item{x}{org object as character vector.}
}
\value{
the days on a task in an org file as a numeric vector.
}
\description{
Do not confuse the number of days on a task with the total time
spent on a task in days. The latter can be calculated from
\\code{extract_time_spent}.
}
\details{
Org mode uses the standard ISO notation for dates and times as it
is defined in ISO 8601. The format is set in
\code{org-time-stamp-custom-formats}. If this variable is
modified, the extraction is likely to fail.
}
\examples{
system.file("extdata", "sample.org", package = "orgclockr") \%>\%
readLines() \%>\%
extract_days_on_task()
##  [1] 0 0 2 0 0 1 5 0 1 2 1 2
}

