% Generated by roxygen2 (4.0.1): do not edit by hand
\name{extract_deadlines}
\alias{extract_deadlines}
\title{Extract deadlines of headings in an org file.}
\usage{
extract_deadlines(x, pattern = "^\\\\ {0,}DEADLINE:\\\\ +<[[:digit:]].*>$")
}
\arguments{
\item{x}{org object as character vector.}

\item{pattern}{a regular expression to match the deadlines set in
the org file.}
}
\value{
the deadlines of an org file as a character vector.
}
\description{
This function extracts the timestamps in lines that match the
\code{pattern}. According to \code{org-deadline-string} these
lines should start with "DEADLINE:". The returned value is in the
format \code{\%Y-\%m-\%d}, so hours and minutes will be omitted.
}
\details{
Org mode uses the standard ISO notation for dates and times as it
is defined in ISO 8601. The format is set in
\code{org-time-stamp-custom-formats}. If this variable is
modified, the extraction is likely to fail.
}
\examples{
system.file("extdata", "sample.org", package = "orgclockr") \%>\%
readLines() \%>\%
extract_deadlines()
##  [1] NA           NA           NA           NA           NA
##  [6] "2014-01-04" NA           NA           NA           NA
## [11] NA           NA
}

