% Generated by roxygen2 (4.0.1): do not edit by hand
\name{extract_intervals}
\alias{extract_intervals}
\title{Extract clock intervals of headings in an org file.}
\usage{
extract_intervals(x, units = "mins")
}
\arguments{
\item{x}{org object as character vector.}

\item{units}{unit of time used.}
}
\value{
a list of double values in the unit of time given.
}
\description{
This expects an even number of timestamps per headline, so in org
files with running clocks the lonely timestamp will be ignored.
}
\examples{
system.file("extdata", "sample.org", package = "orgclockr") \%>\%
readLines() \%>\%
extract_intervals() \%>\%
last()
## [1]  68  17   1  62  16 170
}
\seealso{
\code{extract_time_spent}, \code{extract_days_on_task},
\code{extract_levels}, \code{extract_efforts},
\code{extract_todostates}, \code{extract_headlines},
\code{extract_timestamps}, \code{extract_tags} and
\code{extract_categories} to extract other org elements.
}

