% Generated by roxygen2 (4.0.1): do not edit by hand
\name{extract_levels}
\alias{extract_levels}
\title{Extract levels of headings in an org file.}
\usage{
extract_levels(x)
}
\arguments{
\item{x}{org object as character vector.}
}
\value{
the levels of headlines in an orgfile as an integer
vector.
}
\description{
Get the \code{org-level} in an org file by counting the prepending
heading star characters.
}
\examples{
system.file("extdata", "sample.org", package = "orgclockr") \%>\%
readLines() \%>\%
extract_levels()
##  [1] 1 2 2 2 1 2 2 1 2 2 2
}
\seealso{
\code{extract_todostates}, \code{extract_headlines},
\code{extract_timestamps}, \code{extract_tags} and
\code{extract_categories} to extract other org elements.
}

