% Generated by roxygen2 (4.0.1): do not edit by hand
\name{extract_time_spent}
\alias{extract_time_spent}
\title{Extract the time spent per headline in an org file.}
\usage{
extract_time_spent(x, units = "mins")
}
\arguments{
\item{x}{org object as character vector.}

\item{units}{unit of time used.}
}
\value{
a vector of doubles in the timeformat given by the
according parameter.
}
\description{
This function calculates the difference for each pair of clocking
timestamps and returns the sum of differences per headline. This
expects an even number of timestamps per headline, so in org files
with running clocks the lonely timestamp will be ignored.
}
\examples{
system.file("extdata", "sample.org", package = "orgclockr") \%>\%
readLines() \%>\%
extract_time_spent()
##  [1]   0   0  21   0   0   2 232   0 122 152   2 334
}
\seealso{
\code{extract_intervals}, \code{extract_days_on_task},
\code{extract_levels}, \code{extract_efforts},
\code{extract_todostates}, \code{extract_headlines},
\code{extract_timestamps}, \code{extract_tags} and
\code{extract_categories} to extract other org elements.
}

