% Generated by roxygen2 (4.0.1): do not edit by hand
\name{extract_timestamps}
\alias{extract_timestamps}
\title{Extract the timestamps of an org file.}
\usage{
extract_timestamps(orgfile)
}
\arguments{
\item{orgfile}{org object as character vector.}
}
\value{
the timestamps of an orgfile as a character matrix.
}
\description{
This function is used to extract the timestamps matching the
regexp pattern.
}
\seealso{
\code{extract_headlines}, \code{extract_tags}
and \code{extract_categories} for extractions of other org elements.
}

